requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pyjwt
alembic
pydantic[email]
pydantic-settings
python-jose
alembic
python-multipart